Keep test worker alive until explicitly freed
This change ensures that all messages from the worker are received by
the main process before the worker exits.

It solves issues on Windows that are occurring with Node.js v20 and later
when `workerThreads: false` is set in the AVA configuration.

Fixes: https://github.com/avajs/ava/issues/3390

* Update comment and rename message

Whether the worker exits depends on whether user code is keeping the event loop busy. 'Freed' is a better term.

---------

Co-authored-by: Mark Wubben <mark@novemberborn.net>