Adjust release process
Rulesets prohibit pushing commits on `main` without a PR. This means that running `npm version` within `main` actually
creates more work. Further, the release workflow requires CI to have passed on the tagged commit, but it's triggered
when the tag is pushed.

Simplify by triggering on pushes to `main`, when that commit has a release tag. We can then create a PR for a release
branch, push the tag commit, and trigger the release when merging the commit into `main`. This still has to happen
locally since we must include the tagged commit.